Crate problem086

Problem 86 (Cuboid route)

Solution detail

Supposing that a given cuboid has side lengths x ≤ y≤ z, then the shortest path between opposite corners is equal to sqrt((x + y)^2 + z^2), by "unfolding" the cuboid along its longest sides.

We can use the following process to find the number of cuboids with integer shortest path and maximum side-length equal to m:

Iterate over each k = m + 1, m + 2, m + 3...,

Let d = k^2 - m^2. If D > 4m^2 then stop processing. Otherwise, let s = sqrt(d) and and count how many 1 ≤ x ≤ y ≤ m there are with x + y = s. Add this count the the total number of solutions found so far.

If we finish processing a particular m, and the solution count is now large enough, then we have found our answer.

Functions

answer
Solve the problem, returning the answer as a String

cuboids_with_longest_side
Find the number of cuboids with an integer shortest path and longest side equal to m.

main

pairs_with_sum
Find the number of pairs x, y with 1 ≤ x ≤ y ≤ lim and x + y = d

solve
Find the least value of m such that there are at least the given number of cuboids, with longest side not exceeding m, with an integer shortest path.